LLMTeacher/requirements.txt
FastAPI
uvicorn
python-dotenv
pydantic
httpx
requests
python-multipart 